# release-23.2: rttanalysis, catalog: add benchmark for scanning comments, database-aware scans for comments