Backport 2/2 commits from #114993.

/cc @cockroachdb/release

---

### rttanalysis: add a benchmark for scanning comments

This patch adds a query made by liquibase migrations
as a benchmark test. The example setup (15 dbs, 40 tables each)
mimicks a pattern we have seen before that causes said query to
be particularly slow during the scanning of comments. In addition
to the multiple database setup, this query is benchmarked under
a single database setup as well - allowing us to compare the
behaviors of both patterns.

Epic: none
Informs: https://github.com/cockroachdb/cockroach/issues/114822

Release note: None

---

### catalog: perform database-aware scans for comments in namespace

Previously, we would look at the entirety of system.namespace
when we performed a scan of all comments (even though we only wanted to
pull comments for the current database). We saw that this operation
proved to be slow in cases where we had multiple databases with
multiple tables each. This patch ensures that we only perform a scan
for comments in the current database.

Benchmarks from multiple database test:
```
ORMQueries/liquibase_migrations_on_multiple_dbs-10     sec/op             sec/op vs base
                    				       163.03m ± 41%      22.50m ± 2%   -86.20% (p=0.000 n=10)

ORMQueries/liquibase_migrations_on_multiple_dbs-10     roundtrips         roundtrips vs base
			  			       7.000   ± 0% 	  8.000 ± 0%    +14.29% (p=0.000 n=10)

ORMQueries/liquibase_migrations_on_multiple_dbs-10     B/op               B/op vs base
			  			       44.82Mi ± 11%      27.47Mi ± 6%  -38.71% (p=0.000 n=10)

ORMQueries/liquibase_migrations_on_multiple_dbs-10     allocs/op          allocs/op vs base
			 			       407.3k  ± 15%      253.7k ± 4%   -37.71% (p=0.000 n=10)
```

Benchmarks from single database test:
```
ORMQueries/liquibase_migrations-10     sec/op             sec/op vs base
                                       9.86m    ± 1%      14.29m ± 4%   -28.06% (p=0.000 n=10)

ORMQueries/liquibase_migrations-10     roundtrips         roundtrips vs base
                                       3.000    ± 0%      4.000 ± 0%    +33.33% (p=0.000 n=10)

ORMQueries/liquibase_migrations-10     B/op               B/op vs base
                                       11.739Mi ± 2%      7.606Mi ± 5%  -35.20% (p=0.000 n=10)

ORMQueries/liquibase_migrations-10     allocs/op          allocs/op vs base
			 	       107.51k  ± 4%      67.53k ± 8%   -37.19% (p=0.000 n=10)
```

Epic: none
Fixes: https://github.com/cockroachdb/cockroach/issues/114822

Release note: None


Release justification: Corrects a significant performance regression in getting comments